AVX=32997: Refactor asg and CFT handlerss
